refactor: split out processing
